TimeLib: add partial minutes test

+++ src/main/php/DataSift/Stone/TimeLib/DateInterval.php
@@ -40,7 +40,7 @@ class DateInterval extends \DateInterval
 	 */
 	public function getTotalMinutes()
 	{
-		return $this->getTotalSeconds() / 60;
+		return (int)($this->getTotalSeconds() / 60);
 	}
 
 	/**
